requirements.txt
flask==3.0.0
opencv-python-headless==4.8.1.78
numpy>=1.26.0,<2.0.0
scikit-learn>=1.5.0
Pillow>=10.3.0
flask-cors==4.0.0
gunicorn==21.2.0
werkzeug==3.0.1
python-dotenv==1.0.0
flask-limiter==3.5.0
